DAO-188 Move policy components into components folder
