+++ src/app.tsx
@@ -14,7 +14,7 @@ import NotFoundPage from './pages/not-found';
 import ProposalDetailsPage from './pages/proposal-commons/proposal-details';
 import Proposals from './pages/proposals';
 import Vesting from './pages/vesting';
-import Policies from './pages/policies';
+import Policies from './components/policies';
 import Claims from './pages/claims';
 import ClaimDetails from './pages/claim-details';
 import PolicyDetails from './pages/policy-details';
