+++ src/components/policies/policies.tsx
@@ -1,16 +1,16 @@
 import { FormEventHandler, ReactNode, useMemo } from 'react';
 import { useHistory } from 'react-router';
-import Layout from '../../components/layout';
-import BorderedBox, { Header } from '../../components/bordered-box';
-import RadioButton from '../../components/radio-button';
-import Button from '../../components/button';
-import Input from '../../components/input';
-import CloseIcon from '../../components/icons/close-icon';
-import SearchIcon from '../../components/icons/search-icon';
+import Layout from '../layout';
+import BorderedBox, { Header } from '../bordered-box';
+import RadioButton from '../radio-button';
+import Button from '../button';
+import Input from '../input';
+import CloseIcon from '../icons/close-icon';
+import SearchIcon from '../icons/search-icon';
 import PolicyList from './policy-list';
 import Pagination, { usePagedData } from './pagination';
 import { useQueryParams } from '../../utils';
-import { connectWallet } from '../../components/sign-in/sign-in';
+import { connectWallet } from '../sign-in/sign-in';
 import { useChainData } from '../../chain-data';
 import { useUserPolicies, isActive } from '../../logic/policies';
 import styles from './policies.module.scss';
